+++ obadiah/src/main/java/net/zionsoft/obadiah/notes/NotesActivity.java
@@ -27,6 +27,7 @@ import android.support.v4.util.Pair;
 import android.view.View;
 
 import net.zionsoft.obadiah.R;
+import net.zionsoft.obadiah.biblereading.BibleReadingActivity;
 import net.zionsoft.obadiah.model.analytics.Analytics;
 import net.zionsoft.obadiah.model.datamodel.Settings;
 import net.zionsoft.obadiah.model.domain.Note;
@@ -130,7 +131,7 @@ public class NotesActivity extends BaseRecyclerViewActivity implements NotesView
                     + "-" + item.second.verseIndex.chapter() + "-" + item.second.verseIndex.verse());
             Analytics.logEvent(Analytics.EVENT_SELECT_CONTENT, params);
 
-            finish();
+            startActivity(BibleReadingActivity.newStartIntent(this));
         }
     }
 }